migrate to unity 2017 & update probuilder

+++ Assets/Resources/ProCore/ProBuilder/Shader/pb_VertexPicker.shader
@@ -19,7 +19,7 @@
 		Blend Off
 		Offset -1, -1
 
-		Pass 
+		Pass
 		{
 			Name "Vertices"
 			AlphaTest Greater .25
@@ -74,7 +74,7 @@
 				return o;
 			}
 
-			half4 frag (v2f i) : COLOR
+			float4 frag (v2f i) : COLOR
 			{
 				return i.color;
 			}
